requirements.txt
pandas>=2.0
python-docx>=1.0
python-pptx>=0.6.23
openpyxl>=3.1
pytest>=8.0
pytest-cov>=5.0
